--- app/src/main/java/org/sopt/santamanitto/auth/SignUpRequestModel.kt
@@ -1,6 +1,0 @@
-package org.sopt.santamanitto.auth
-
-data class SignUpRequestModel(
-    val serialNumber: String,
-    val userName: String
-)
