+++ app/src/main/java/org/sopt/santamanitto/auth/data/request/SignInRequestModel.kt
@@ -1,4 +1,4 @@
-package org.sopt.santamanitto.auth
+package org.sopt.santamanitto.auth.data.request
 
 data class SignInRequestModel(
     val serialNumber: String
