+++ app/src/main/java/org/sopt/santamanitto/auth/data/request/SignUpRequestModel.kt
@@ -1,0 +1,6 @@
+package org.sopt.santamanitto.auth.data.request
+
+data class SignUpRequestModel(
+    val serialNumber: String,
+    val name: String
+)
