+++ app/src/main/java/org/sopt/santamanitto/auth/data/response/SignInResponseModel.kt
@@ -1,0 +1,6 @@
+package org.sopt.santamanitto.auth.data.response
+
+data class SignInResponseModel(
+    val accessToken: String,
+    val id: String
+)
