+++ app/src/main/java/org/sopt/santamanitto/auth/data/response/SignUpResponseModel.kt
@@ -1,0 +1,6 @@
+package org.sopt.santamanitto.auth.data.response
+
+data class SignUpResponseModel(
+    val accessToken: String,
+    val id: String
+)
