+++ app/src/main/java/org/sopt/santamanitto/user/signin/viewmodel/ConditionViewModel.kt
@@ -3,7 +3,7 @@ package org.sopt.santamanitto.user.signin.viewmodel
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
-import org.sopt.santamanitto.user.data.UserLoginModel
+import org.sopt.santamanitto.auth.data.response.SignUpResponseModel
 import org.sopt.santamanitto.user.data.controller.UserController
 import org.sopt.santamanitto.user.data.source.UserMetadataSource
 import javax.inject.Inject
@@ -31,24 +31,26 @@ class ConditionViewModel @Inject constructor(
             return
         }
         _isWaitingForResponse = true
-        userController.createAccount(userName, serialNumber, object : UserController.CreateAccountCallback {
-
-            override fun onCreateAccountSuccess(userLoginModel: UserLoginModel) {
-                userMetadataSource.run {
-                    userLoginModel.let {
-                        setUserName(it.userName)
-                        setAccessToken(it.accessToken)
-                        setUserId(it.id)
+        userController.createAccount(
+            userName,
+            serialNumber,
+            object : UserController.CreateAccountCallback {
+                override fun onCreateAccountSuccess(signUpResponseModel: SignUpResponseModel) {
+                    userMetadataSource.run {
+                        signUpResponseModel.let {
+//                            setUserName(it.userName)
+                            setAccessToken(it.accessToken)
+                            setUserId(it.id)
+                        }
                     }
+                    userSaveSuccess.value = true
+                    _isWaitingForResponse = false
                 }
-                userSaveSuccess.value = true
-                _isWaitingForResponse = false
-            }
 
-            override fun onCreateAccountFailed() {
-                userSaveFail.value = true
-                _isWaitingForResponse = false
-            }
-        })
+                override fun onCreateAccountFailed() {
+                    userSaveFail.value = true
+                    _isWaitingForResponse = false
+                }
+            })
     }
 }
